backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
python-multipart==0.0.6
pillow==10.1.0
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.5.0
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
